server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asgiref==3.11.0
attrs==25.4.0
bitarray==3.8.0
bitstring==4.3.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
Flask==3.1.2
frozenlist==1.8.0
idna==3.11
intelhex==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
propcache==0.4.1
pycparser==2.23
Pygments==2.19.2
pyserial==3.5
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
reedsolo==1.7.0
requests==2.32.5
rich==14.2.0
rich-click==1.9.5
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
yarl==1.22.0
